betsson_dw/requirements.txt
pandas==2.0.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
